feat(help): update description
